backend/weather_processor/requirements.txt
functions-framework
google-cloud-storage
google-genai